Initial OpenCL implementation
Adds a rough OpenCL implementation. For now, two separate executables
are build: BitCrack64.exe (CUDA) and clBitCrack64.exe (OpenCL).

The bloom filter size is currently hard-coded.
Performance is about half of what is achieved with the CUDA version.

The OpenCL code lives in CLKeySearchDevice. The *.cl files also
must be present in the same directory as clBitCrack.exe, since it
needs to read and compile them at runtime.