Fixed bug where incorrect results would be produced on AMD devices.
The problem is in the modular inversion (invModP() function of
secp2556k1.cl). There is section that is looped 20 times. Breaking
the loop into two loops fixes the issue. The exact cause is still
unkown, but it seems to be an issue in AMD's OpenCL compiler, as
this does not happen in the Mesa OpenCL compiler, or nVidia's
OpenCL compiler.